client/requirements.txt
Flask==3.1.2
requests==2.32.5
grpcio==1.75.1
grpcio-tools==1.75.1
flasgger==0.9.7.1
python-dotenv==1.0.0
